cpu: add mechanism to enqueue jobs for parallel execution [1/2]
Using the util::Worker, provide a "queue" to schedule a job to run
across multiple threads in a grid. The queue is part of the static
CPU Device class: grid::cpu::Queue.

Use the following method to add a function to the device queue:

  queue.Enqueue(dimensions, sizes, FUNCTION);

The arguments dimensions and sizes define the overall dimensions
and sizes of each "tile".

The FUNCTION is called with the position in the grid.

To synchronize between job, use queue.Sync(), which will wait for all
currently scheduled jobs to complete.

Signed-off-by: Chris Zankel <chris@zankel.net>