tensor: provide an Array specialization for a View
This will enable dropping a separate TensorView class and implement
views just as a "regular" Tensor derived from the Array View.

The first step is to just declare the new Array View specialization and
derive the TensorView class from it.

Also, decleare a new templated structure for Views similar to DeviceMemory,
StaticMemory, etc.

Signed-off-by: Chris Zankel <chris@zankel.net>