tensor: fixup CPU unittest and constructor from TensorView
The Tensor alias template used in the unittest for the CPU was just
passing through the device type from a template instead of defining
it to device::CPU. It then relied on the default values in Tensor's
CDAT rules to use the correct device.

Changing the tensor alias to CPU to align with other implementation
required to also be more specific with TensorView.

Signed-off-by: Chris Zankel <chris@zankel.net>