tensor/unittest: update CTAD rules; alias templates deduct. in OSX
The CTAD rules need to be updated due to the recent changes.

The clang version clang-1700.3.19.1 included in the Xcode 26.0.1,
Build version 17A400 now supports P1814r0:

Wording for Class Template Argument Deduction for Alias Templates
https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2019/p1814r0.html

So, enable it by default for 17.0.3 for Apple.

With this clang update, we will likely drop the alternative declarations
of a helper tensor in unittest in the near future.

Signed-off-by: Chris Zankel <chris@zankel.net>